Refactoring
Separation test utilities and test services
