+++ jcommune-view/jcommune-web-view/src/test/java/org/jtalks/jcommune/test/model/User.groovy
@@ -1,0 +1,29 @@
+/**
+ * Copyright (C) 2011  JTalks.org Team
+ * This library is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU Lesser General Public
+ * License as published by the Free Software Foundation; either
+ * version 2.1 of the License, or (at your option) any later version.
+ * This library is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ * Lesser General Public License for more details.
+ * You should have received a copy of the GNU Lesser General Public
+ * License along with this library; if not, write to the Free Software
+ * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
+ */
+package org.jtalks.jcommune.test.model
+
+import static org.apache.commons.lang.RandomStringUtils.randomAlphabetic
+
+
+/**
+ * @author Mikhail Stryzhonok
+ */
+class User {
+    String username = randomAlphabetic(25)
+    String email = randomAlphabetic(40) + "@sample.ru"
+    String password = randomAlphabetic(50)
+    String confirmation = password
+    String honeypot = ""
+}
