+++ jcommune-view/jcommune-web-view/src/test/java/org/jtalks/jcommune/test/service/BranchService.java
@@ -12,7 +12,7 @@
  * License along with this library; if not, write to the Free Software
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
  */
-package org.jtalks.jcommune.model.utils;
+package org.jtalks.jcommune.test.service;
 
 import org.jtalks.jcommune.model.dao.BranchDao;
 import org.jtalks.jcommune.model.entity.Branch;
@@ -21,9 +21,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 /**
  * @author Mikhail Stryzhonok
  */
-public class Branches {
-
-    private static final Branches INSTANCE = new Branches();
+public class BranchService {
 
     @Autowired
     private BranchDao branchDao;
