+++ jcommune-view/jcommune-web-view/src/test/java/org/jtalks/jcommune/test/service/GroupsService.java
@@ -12,7 +12,7 @@
  * License along with this library; if not, write to the Free Software
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
  */
-package org.jtalks.jcommune.model.utils;
+package org.jtalks.jcommune.test.service;
 
 import org.jtalks.common.model.entity.Group;
 import org.jtalks.jcommune.model.dao.GroupDao;
@@ -22,7 +22,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 /**
  * @author Mikhail Stryzhonok
  */
-public class Groups {
+public class GroupsService {
 
     @Autowired
     private GroupDao groupDao;
