+++ jcommune-view/jcommune-web-view/src/test/java/org/jtalks/jcommune/test/service/PermissionGranter.java
@@ -12,7 +12,7 @@
  * License along with this library; if not, write to the Free Software
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
  */
-package org.jtalks.jcommune.model.utils;
+package org.jtalks.jcommune.test.service;
 
 import org.jtalks.common.model.entity.Entity;
 import org.jtalks.common.model.entity.Group;
