+++ jcommune-view/jcommune-web-view/src/test/java/org/jtalks/jcommune/test/utils/page/PageUsers.groovy
@@ -17,7 +17,7 @@ package org.jtalks.jcommune.test.utils.page;
 import org.jtalks.jcommune.test.utils.Users;
 import org.jtalks.jcommune.test.utils.exceptions.ValidationException;
 import org.jtalks.jcommune.test.utils.exceptions.WrongResponseException;
-import org.jtalks.jcommune.test.utils.model.User;
+import org.jtalks.jcommune.test.model.User;
 import org.jtalks.jcommune.web.controller.UserController;
 import org.springframework.test.web.servlet.MvcResult;
 import org.springframework.validation.BindingResult;
